+++ src/synthesizer/__init__.py
@@ -1,24 +1,17 @@
+# Import the various utils submodules to make them available at the top level
+from synthesizer.utils import art
+from synthesizer.utils import integrate
+from synthesizer.utils import plt
+from synthesizer.utils import stats
+from synthesizer.utils import util_funcs
+
 # Import the main classes to make them available at the top level
-from synthesizer import abundances
 from synthesizer.grid import Grid
 
 
 # Import an alais the galaxy factory function
 from synthesizer.galaxy import galaxy
 from synthesizer.galaxy import galaxy as Galaxy
-
-
-# Import the various utils submodules to make them available at the top level
-from synthesizer.utils import art
-from synthesizer.utils import conversions
-from synthesizer.utils import ascii_table
-from synthesizer.utils import exceptions
-from synthesizer.utils import integrate
-from synthesizer.utils import plt
-from synthesizer.utils import stats
-from synthesizer.utils import util_funcs
-from synthesizer.utils import units
-from synthesizer.utils import warnings
 
 
 # Import the various extensions submodules to make them available
