+++ src/synthesizer/emission_models/attenuation/__init__.py
@@ -1,2 +1,2 @@
 # Unpack the attenuation laws into the parent package
-from synthesizer.emission_models.attenuation_laws import *
+from synthesizer.emission_models.attenuation.attenuation_laws import *
