+++ src/synthesizer/emission_models/attenuation/attenuation_laws.py
