+++ src/synthesizer/exceptions.py
