+++ src/synthesizer/parametric/sf_hist.py
@@ -21,7 +21,7 @@ import numpy as np
 from unyt import yr
 
 from synthesizer import exceptions
-from synthesizer.stats import weighted_mean, weighted_median
+from synthesizer.utils.stats import weighted_mean, weighted_median
 
 # Define a list of the available parametrisations
 parametrisations = (
