+++ src/synthesizer/utils/__init__.py
@@ -1,2 +1,5 @@
 # Import all functions into the top level namespace
 from synthesizer.utils.util_funcs import *
+
+# Import the table formatter for __str__ methods
+from synthesizer.utils.ascii_table import TableFormatter
